Move task runner to shared.
